update REST endpoints to utilize user information

+++ src/routes/centerline/[id]/+server.ts
@@ -1,18 +1,28 @@
 import prisma from '$lib/server/prisma';
 import type { RequestHandler } from './$types';
-import { json } from '@sveltejs/kit';
+import { error, json } from '@sveltejs/kit';
 
 /** Get specific Centerline */
-export const GET: RequestHandler = async ({ params }) => {
-	return json(
-		await prisma.centerline.findUnique({
-			where: { id: Number(params.id) },
-			include: { markers: true }
-		})
-	);
+export const GET: RequestHandler = async ({ params, locals }) => {
+	const { user } = await locals.auth.validateUser();
+	if (!user) throw error(401, 'Must be signed in to access this endpoint');
+
+	const item = await prisma.centerline.findUnique({
+		where: { id: Number(params.id) },
+		include: { markers: true }
+	});
+	if (item?.user_id !== user.userId) throw error(403, 'Cannot access this resource');
+
+	return json(item);
 };
 
 /** Delete Centerline */
-export const DELETE: RequestHandler = async ({ params }) => {
+export const DELETE: RequestHandler = async ({ params, locals }) => {
+	const { user } = await locals.auth.validateUser();
+	if (!user) throw error(401, 'Must be signed in to access this endpoint');
+
+	const item = await prisma.centerline.findUnique({ where: { id: Number(params.id) } });
+	if (item?.user_id !== user.userId) throw error(403, 'Cannot access this resource');
+
 	return json(await prisma.centerline.delete({ where: { id: Number(params.id) } }));
 };
